Changes to be committed: 	modified:   syntaxes/ps.tmLanguage.json
Changed some of the naming around so the pine-specific rules could
be easily identified and highlighted.

The problem is this breaks all the other themes besides the
new one. So, the rest could be overhauled just as well.
Would just be a matter of find/replace-all for the mismatches
in the other themese.
But the layout of the rules is considerably easier to follow.

This also adds multiple (about 20) built-in functions and variables
that were added since the last release.